+++ config/test/mu2edaq22.C
@@ -4,20 +4,26 @@
 // init_run_configuration : the name and the call signature are is fixed 
 //                          and can't be changed
 ///////////////////////////////////////////////////////////////////////////////
-#include "otsdaq-mu2e-tracker/gui/DtcGui.hh"
+#include "otsdaq-mu2e-tracker/Gui/DtcGui.hh"
 
-int init_run_configuration(DtcGui::DtcData_t* DtcData) {
+int init_run_configuration(DtcGui* X) {
   int rc(0);
 
   printf("[init_run_configuration] : test : DtcGui : %s\n",gSystem->Getenv("HOSTNAME"));
 
-  DtcData[0].fName     = "DTC";
-  DtcData[0].fPcieAddr = 0;
-  DtcData[0].fLinkMask = 0x1;  // 
+  DtcGui::DtcData_t* dat = (DtcGui::DtcData_t*) X->fDtcData;
 
-  DtcData[1].fName     = "DTC";
-  DtcData[1].fPcieAddr = 1;
-  DtcData[1].fLinkMask = 0x1;  // 
+  X->fNDtcs           = 2;              // installed on a machine
+
+  dat[0].fName        = "DTC";
+  dat[0].fPcieAddr    = 0;
+  dat[0].fLinkMask    = 0x1;            // 
+  dat[0].fNLinkedDtcs = 0;              // irrelevant for a DTC
+
+  dat[1].fName        = "DTC";
+  dat[1].fPcieAddr    = 1;
+  dat[1].fLinkMask    = 0x1;  // 
+  dat[1].fNLinkedDtcs = 2;
 
   return rc;
 }
