+++ otsdaq-mu2e-tracker/Gui/DtcGui_gui_CfoTab.cc
@@ -1,11 +1,11 @@
 
-#include "otsdaq-mu2e-tracker/gui/DtcGui.hh"
+#include "otsdaq-mu2e-tracker/Gui/DtcGui.hh"
 
 using namespace trkdaq;
 using namespace std;
 
 //-----------------------------------------------------------------------------
-void DtcGui::BuildDtcTabElement(TGTab*& Tab, DtcTabElement_t& DtcTel, DtcData_t* DtcData) {
+void DtcGui::BuildCfoTabElement(TGTab*& Tab, DtcTabElement_t& DtcTel, DtcData_t* DtcData) {
 
   DtcTel.fData = DtcData;
 
@@ -22,33 +22,7 @@ void DtcGui::BuildDtcTabElement(TGTab*& Tab, DtcTabElement_t& DtcTel, DtcData_t*
 //-----------------------------------------------------------------------------
 // DTC's have ROCs, CFO's - do not
 //-----------------------------------------------------------------------------
-  int id = 0;
-
-  if (DtcData->fName == "DTC") {
-    DtcTel.fRocTab = new TGTab(group,10,10);
-    group->AddFrame(DtcTel.fRocTab, new TGLayoutHints(kLHintsLeft | kLHintsTop,2,2,2,2));
-
-    for (int i=0; i<6; i++) {
-      BuildRocTabElement(DtcTel.fRocTab,DtcTel.fRocTel[i], &fDtcData->fRocData[i]);
-    }
-//-----------------------------------------------------------------------------
-// position the ROC tab inside the DTC tab
-//-----------------------------------------------------------------------------
-    DtcTel.fRocTab->MoveResize(520,20,360,190);
-    DtcTel.fRocTab->Connect("Selected(Int_t)", "DtcGui", this, "DoRocTab(Int_t)");
-
-    DtcTel.fRocTab->SetTab(id);
-
-//-----------------------------------------------------------------------------
-// set active ROC tab
-//-----------------------------------------------------------------------------
-    DtcTel.fActiveRocID  = id;
-
-    DtcTel.fActiveRocTel = &DtcTel.fRocTel[id];
-    DtcTel.fRocTabColor  = DtcTel.fActiveRocTel->fTab->GetBackground();
-    
-    DtcTel.fActiveRocTel->fTab->ChangeBackground(fYellow);
-  }
+//  int id = 0;
 //------------------------------------------------------------------------------
 // a) graphics context changes
 //-----------------------------------------------------------------------------
@@ -85,7 +59,7 @@ void DtcGui::BuildDtcTabElement(TGTab*& Tab, DtcTabElement_t& DtcTel, DtcData_t*
   tb->Connect("Pressed()", "DtcGui", this, "print_dtc_status()");
   tb->ChangeBackground(fValidatedColor);
 //-----------------------------------------------------------------------------
-// DTC write register 
+// write register 
 //-----------------------------------------------------------------------------
   tb = new TGTextButton(group,"writeR",-1,TGTextButton::GetDefaultGC()(),
                         TGTextButton::GetDefaultFontStruct(),kRaisedFrame);
@@ -99,7 +73,7 @@ void DtcGui::BuildDtcTabElement(TGTab*& Tab, DtcTabElement_t& DtcTel, DtcData_t*
   tb->Connect("Pressed()", "DtcGui", this, "write_dtc_register()");
   tb->ChangeBackground(fValidatedColor);
 //-----------------------------------------------------------------------------
-// column 1 raw 3: DTC read register 
+// column 1 raw 3: read register 
 //-----------------------------------------------------------------------------
   tb = new TGTextButton(group,"readR",-1,TGTextButton::GetDefaultGC()(),
                         TGTextButton::GetDefaultFontStruct(),kRaisedFrame);
@@ -113,7 +87,7 @@ void DtcGui::BuildDtcTabElement(TGTab*& Tab, DtcTabElement_t& DtcTel, DtcData_t*
   tb->Connect("Pressed()", "DtcGui", this, "read_dtc_register()");
   tb->ChangeBackground(fValidatedColor);
 //-----------------------------------------------------------------------------
-// column 1 raw 4: DTC soft reset 
+// column 1 raw 4: soft reset 
 //-----------------------------------------------------------------------------
   tb = new TGTextButton(group,"soft reset",-1,TGTextButton::GetDefaultGC()(),
                         TGTextButton::GetDefaultFontStruct(),kRaisedFrame);
@@ -127,7 +101,7 @@ void DtcGui::BuildDtcTabElement(TGTab*& Tab, DtcTabElement_t& DtcTel, DtcData_t*
   tb->Connect("Pressed()", "DtcGui", this, "dtc_soft_reset()");
   tb->ChangeBackground(fValidatedColor);
 //-----------------------------------------------------------------------------
-// column 1 raw 5: DTC hard reset 
+// column 1 raw 5: hard reset 
 //-----------------------------------------------------------------------------
   tb = new TGTextButton(group,"hard reset",-1,TGTextButton::GetDefaultGC()(),
                         TGTextButton::GetDefaultFontStruct(),kRaisedFrame);
@@ -172,53 +146,20 @@ void DtcGui::BuildDtcTabElement(TGTab*& Tab, DtcTabElement_t& DtcTel, DtcData_t*
   rr->SetText("0x9100");
   rr->MoveResize(x0+dx+10,y0+2*(dy+5),dx2,dy);
   DtcTel.fRegR = rr;
-
-  if (DtcData->fName == "DTC") {
-//-----------------------------------------------------------------------------
-// column 2 raw 4 : print DTC Firefly temp, different for CFO
-//-----------------------------------------------------------------------------
-    tb = new TGTextButton(group,"FF temp",-1,TGTextButton::GetDefaultGC()(),
-                          TGTextButton::GetDefaultFontStruct(),kRaisedFrame);
-    group->AddFrame(tb, new TGLayoutHints(kLHintsLeft | kLHintsTop,2,2,2,2));
-
-    tb->SetTextJustify(36);
-    tb->SetMargins(0,0,0,0);
-    tb->SetWrapLength(-1);
-  
-    tb->MoveResize(x0+dx+10,y0+(dy+5)*3,dx2,dy);
-    tb->Connect("Pressed()", "DtcGui", this, "print_firefly_temp()");
-    tb->ChangeBackground(fValidatedColor);
-//-----------------------------------------------------------------------------
-// column 2 raw 5 : init readout mode with the external CFO
-//-----------------------------------------------------------------------------
-    tb = new TGTextButton(group,"init ext CFO",-1,TGTextButton::GetDefaultGC()(),
-                          TGTextButton::GetDefaultFontStruct(),kRaisedFrame);
-    group->AddFrame(tb, new TGLayoutHints(kLHintsLeft | kLHintsTop,2,2,2,2));
-
-    tb->SetTextJustify(36);
-    tb->SetMargins(0,0,0,0);
-    tb->SetWrapLength(-1);
-  
-    tb->MoveResize(x0+dx+10,y0+(dy+5)*4,dx2,dy);
-    tb->Connect("Pressed()", "DtcGui", this, "init_external_cfo_readout_mode()");
-    tb->ChangeBackground(fValidatedColor);
-  }
-  else if (DtcData->fName == "CFO") {
 //-----------------------------------------------------------------------------
 // column 2 raw 4 : launch current run plan
 //-----------------------------------------------------------------------------
-    tb = new TGTextButton(group,"launch",-1,TGTextButton::GetDefaultGC()(),
-                          TGTextButton::GetDefaultFontStruct(),kRaisedFrame);
-    group->AddFrame(tb, new TGLayoutHints(kLHintsLeft | kLHintsTop,2,2,2,2));
-
-    tb->SetTextJustify(36);
-    tb->SetMargins(0,0,0,0);
-    tb->SetWrapLength(-1);
+  tb = new TGTextButton(group,"launch",-1,TGTextButton::GetDefaultGC()(),
+                        TGTextButton::GetDefaultFontStruct(),kRaisedFrame);
+  group->AddFrame(tb, new TGLayoutHints(kLHintsLeft | kLHintsTop,2,2,2,2));
   
-    tb->MoveResize(x0+dx+10,y0+(dy+5)*3,dx2,dy);
-    tb->Connect("Pressed()", "DtcGui", this, "cfo_launch_run_plan()");
-    tb->ChangeBackground(fValidatedColor);
-  }
+  tb->SetTextJustify(36);
+  tb->SetMargins(0,0,0,0);
+  tb->SetWrapLength(-1);
+  
+  tb->MoveResize(x0+dx+10,y0+(dy+5)*3,dx2,dy);
+  tb->Connect("Pressed()", "DtcGui", this, "cfo_launch_run_plan()");
+  tb->ChangeBackground(fValidatedColor);
 //-----------------------------------------------------------------------------
 // column 3 row 1: write value: 1. label , 2: entry field  3: label
 //-----------------------------------------------------------------------------
@@ -242,7 +183,7 @@ void DtcGui::BuildDtcTabElement(TGTab*& Tab, DtcTabElement_t& DtcTel, DtcData_t*
   wval->MoveResize(x0+dx+10+dx2+10,y0+(dy+5) ,dx3,dy);
   DtcTel.fValW = wval;
 //-----------------------------------------------------------------------------
-// column 3 row 3: label
+// column 3 row 3: label (value of the read register)
 //-----------------------------------------------------------------------------
   lab = new TGLabel(group,"0x00000000");
   group->AddFrame(lab, new TGLayoutHints(kLHintsLeft | kLHintsTop,2,2,2,2));
@@ -251,20 +192,86 @@ void DtcGui::BuildDtcTabElement(TGTab*& Tab, DtcTabElement_t& DtcTel, DtcData_t*
   lab->SetWrapLength(-1);
   lab->MoveResize(x0+dx+10+dx2+10,y0+2*(dy+5),dx3,dy);
   DtcTel.fValR = lab;
+//-----------------------------------------------------------------------------
+// column 4 row 1: label: CFO timing chain link 
+//-----------------------------------------------------------------------------
+  int c4_dx = x0+dx+10+dx2+10+dx3+10;
 
-  if (DtcData->fName == "DTC") {
-    tb = new TGTextButton(group,"read D",-1,TGTextButton::GetDefaultGC()(),
-                          TGTextButton::GetDefaultFontStruct(),kRaisedFrame);
-    group->AddFrame(tb, new TGLayoutHints(kLHintsLeft | kLHintsTop,2,2,2,2));
+  lab = new TGLabel(group,"CFO link");
+  group->AddFrame(lab, new TGLayoutHints(kLHintsLeft | kLHintsTop,2,2,2,2));
+  lab->SetTextJustify(36);
+  lab->SetMargins(0,0,0,0);
+  lab->SetWrapLength(-1);
+  lab->MoveResize(c4_dx,y0,dx3,dy);
+  //  DtcTel.fCfoLink = lab;
+//-----------------------------------------------------------------------------
+// column 4 row 2: label: N(DTCs) in the timing chain 
+//-----------------------------------------------------------------------------
+  lab = new TGLabel(group,"N(DTCs)");
+  group->AddFrame(lab, new TGLayoutHints(kLHintsLeft | kLHintsTop,2,2,2,2));
+  lab->SetTextJustify(36);
+  lab->SetMargins(0,0,0,0);
+  lab->SetWrapLength(-1);
+  lab->MoveResize(c4_dx,y0+(dy+5),dx3,dy);
+  //  DtcTel.fNDtcs = lab;
+//-----------------------------------------------------------------------------
+// column 4 row 3: label: run plan
+//-----------------------------------------------------------------------------
+  lab = new TGLabel(group,"Run Plan");
+  group->AddFrame(lab, new TGLayoutHints(kLHintsLeft | kLHintsTop,2,2,2,2));
+  lab->SetTextJustify(36);
+  lab->SetMargins(0,0,0,0);
+  lab->SetWrapLength(-1);
+  lab->MoveResize(c4_dx,y0+2*(dy+5),dx3,dy);
+//-----------------------------------------------------------------------------
+// column 5 row 1: input field: CFO timing chain link 
+//-----------------------------------------------------------------------------
+  int c5_dx = c4_dx+dx3+10;
 
-    tb->SetTextJustify(36);
-    tb->SetMargins(0,0,0,0);
-    tb->SetWrapLength(-1);
+  rr = new TGTextEntry(group, new TGTextBuffer(14),-1,uGC->GetGC(),
+                       ufont->GetFontStruct(),kSunkenFrame | kOwnBackground);
+  group->AddFrame(rr, new TGLayoutHints(kLHintsLeft | kLHintsTop,2,2,2,2));
+  rr->SetMaxLength(4096);
+  rr->SetAlignment(kTextLeft);
+  rr->SetText("1");
+  rr->MoveResize(c5_dx,y0,dx3,dy);
+  DtcTel.fTimeChainLink = rr;
+//-----------------------------------------------------------------------------
+// column 5 row 2: input field: N(DTCs)
+//-----------------------------------------------------------------------------
+  rr = new TGTextEntry(group, new TGTextBuffer(14),-1,uGC->GetGC(),
+                       ufont->GetFontStruct(),kSunkenFrame | kOwnBackground);
+  group->AddFrame(rr, new TGLayoutHints(kLHintsLeft | kLHintsTop,2,2,2,2));
+  rr->SetMaxLength(4096);
+  rr->SetAlignment(kTextLeft);
+  rr->SetText("1");
+  rr->MoveResize(c5_dx,y0+(dy+5),dx3,dy);
+  DtcTel.fNDtcs = rr;
+//-----------------------------------------------------------------------------
+// column 5 row 3: input field: N(DTCs)
+//-----------------------------------------------------------------------------
+  rr = new TGTextEntry(group, new TGTextBuffer(14),-1,uGC->GetGC(),
+                       ufont->GetFontStruct(),kSunkenFrame | kOwnBackground);
+  group->AddFrame(rr, new TGLayoutHints(kLHintsLeft | kLHintsTop,2,2,2,2));
+  rr->SetMaxLength(4096);
+  rr->SetAlignment(kTextLeft);
+  rr->SetText("run_066.bin");
+  rr->MoveResize(c5_dx,y0+2*(dy+5),dx3,dy);
+  DtcTel.fRunPlan = rr;
+//-----------------------------------------------------------------------------
+// column (4+5) raw 4 : "Init Run Plan" for given CFO time link and N(DTCs
+//-----------------------------------------------------------------------------
+  tb = new TGTextButton(group,"init readout",-1,TGTextButton::GetDefaultGC()(),
+                        TGTextButton::GetDefaultFontStruct(),kRaisedFrame);
+  group->AddFrame(tb, new TGLayoutHints(kLHintsLeft | kLHintsTop,2,2,2,2));
   
-    tb->MoveResize(x0+dx+10+dx2+10,y0+(dy+5)*3,dx3,dy);
-    tb->Connect("Pressed()", "DtcGui", this, "read_subevents()");
-    // tb->ChangeBackground(fValidatedColor);  // not yet
-  }
+  tb->SetTextJustify(36);
+  tb->SetMargins(0,0,0,0);
+  tb->SetWrapLength(-1);
+  
+  tb->MoveResize(c4_dx,y0+3*(dy+5),dx3+10+dx3,dy);
+  tb->Connect("Pressed()", "DtcGui", this, "cfo_init_readout()");
+  // tb->ChangeBackground(fValidatedColor);
 //-----------------------------------------------------------------------------
 // resize the DTC group panel
 //-----------------------------------------------------------------------------
