+++ otsdaq-mu2e-tracker/Gui/classes.h
@@ -1,0 +1,4 @@
+#include <memory> // std::shared_ptr, std::make_shared
+#include <cstring> // memcpy
+
+#include "otsdaq-mu2e-tracker/Gui/DtcGui.hh"
