+++ otsdaq-mu2e-tracker/Gui/root/CMakeLists.txt
@@ -3,13 +3,15 @@ include(ArtDictionary)
 
 art_dictionary(
   DICTIONARY_LIBRARIES
-  otsdaq_mu2e_tracker::otsdaq-mu2e-tracker_ui
-  artdaq_core_mu2e::artdaq-core-mu2e_Overlays
-  mu2e_pcie_utils::CFOInterface
-  mu2e_pcie_utils::DTCInterface
+  otsdaq-mu2e-tracker_gui
+#  artdaq_core_mu2e::artdaq-core-mu2e_Overlays
+#  mu2e_pcie_utils::CFOInterface
+#  mu2e_pcie_utils::DTCInterface
 )
 
-get_target_property(DictCompileOptions otsdaq-mu2e-tracker_ui_root_dict COMPILE_OPTIONS)
+set (dict_name "otsdaq-mu2e-tracker_gui_root_dict")
+
+get_target_property(DictCompileOptions ${dict_name} COMPILE_OPTIONS)
 
 if(DictCompileOptions)
 	string(REPLACE "-fsanitize=thread" "" DictCompileOptions "${DictCompileOptions}")
@@ -17,6 +19,6 @@ if(DictCompileOptions)
 	string(REPLACE "-ltsan" "" DictCompileOptions "${DictCompileOptions}")
 	string(REPLACE "-lasan" "" DictCompileOptions "${DictCompileOptions}")
 
-	set_target_properties(otsdaq-mu2e-tracker_ui_root_dict PROPERTIES COMPILE_OPTIONS "${DictCompileOptions}")
+	set_target_properties(${dict_name} PROPERTIES COMPILE_OPTIONS "${DictCompileOptions}")
 endif()
