+++ otsdaq-mu2e-tracker/Gui/root/classes_def.xml
@@ -1,0 +1,5 @@
+<lcgdict>
+
+  <class name="DtcGui" />
+
+</lcgdict>
