+++ otsdaq-mu2e-tracker/Ui/classes.h
@@ -1,0 +1,5 @@
+#include <memory> // std::shared_ptr, std::make_shared
+#include <cstring> // memcpy
+
+#include "otsdaq-mu2e-tracker/Ui/CfoInterface.hh"
+#include "otsdaq-mu2e-tracker/Ui/DtcInterface.hh"
