+++ otsdaq-mu2e-tracker/Ui/root/classes_def.xml
@@ -1,0 +1,6 @@
+<lcgdict>
+
+  <class name="trkdaq::CfoInterface" />
+  <class name="trkdaq::DtcInterface" />
+
+</lcgdict>
